requirements.txt
requests[security]
pronouncing
lxml
beautifulsoup4
pymongo
nltk
-e git+https://github.com/datascopeanalytics/traces.git#egg=traces
spacy
PyWavelets
ipdb
pyenchant
unidecode
termcolor
